Further work on dexvert 1.0.0 release. AVX support with a portable CPU and tensor in docker is being a pita

+++ release/dexserver.sh
@@ -4,19 +4,18 @@
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
-scratch=400
+scratch=100
 ram=$(awk '/MemTotal/{print int($2*0.7/1024)}' /proc/meminfo)
 tmpDir="/tmp"
 cores=$(( $(nproc) * 95 / 100 ))
 admin=0
 
 # TODO: Add support for changing IP address assigned to VM
-# TODO: Add support for changing cpu type, as for some systems the default 'host' might not work (for example tensor requires avx)
 # TODO: Add support for a publicHost and publicPort option that will tunnel from public to the VM's ip for remote ssh access
 
 show_usage() {
   echo "Usage: $0 [--scratch=<size>] [--ram=<size>] [--scratchDir=<dir>] [--cores=<num>] [--help]"
-  echo "  --scratch=<size>     Set the size of the scratch disk in gigabytes (default: 400GB)."
+  echo "  --scratch=<size>     Set the size of the scratch disk in gigabytes (default: 100GB). This is where output files are placed while being processed."
   echo "  --ram=<size>         Set the size of RAM in megabytes (eg 100 for 100MB) or 0.7 for 70% of RAM (default 0.8 or 80% of total system RAM)."
   echo "  --tmpDir=<dir>       Set the tmp directory to use (default: /tmp)."
   echo "  --cores=<num>        Set the number of cores to use. Either # like 30 for 30 cores or 0.7 for 70% of cores (default 0.95 or 95%)."
@@ -78,6 +77,12 @@ fi
 scratchDir="$tmpDir/$(generate_random_string)"
 mkdir -p "$scratchDir"
 
+scratchAvailable=$(command df -BG --output=avail "$tmpDir" | tail -n1 | tr -d 'G ')
+if [ "$scratchAvailable" -lt "$scratch" ]; then
+	echo "Not enough space in $tmpDir to create scratch disk. Only $scratchAvailable GB available, but --scratch is set to $scratch GB."
+	exit
+fi
+
 echo "Creating scratch.qcow2 file in $scratchDir with size ${scratch}G..."
 scratchFilePath="$scratchDir/scratch.qcow2"
 rm -f "$scratchFilePath"
@@ -88,7 +93,7 @@ hdFilePath="$SCRIPT_DIR/hd.qcow2"
 echo "Launching qemu-system-x86_64 with RAM size $ram MB, using $cores cores, and scratch directory $scratchDir..."
 if [ "$admin" -eq 1 ]; then
 	echo "Launching in ADMIN mode..."
-	qemu-system-x86_64 -monitor telnet::47023,server,nowait -machine accel=kvm,dump-guest-core=off -cpu host -vga std -drive format=qcow2,file="$hdFilePath",if=virtio -device virtio-rng-pci -m size="$ram"M -smp "cores=$cores" -boot order=c -netdev user,net=192.168.47.0/24,dhcpstart=192.168.47.47,hostfwd=tcp:127.0.0.1:47022-192.168.47.47:22,id=nd1 -device virtio-net,netdev=nd1 -drive format=qcow2,file="$scratchFilePath",if=virtio -drive format=qcow2,file="$SCRIPT_DIR/admin.qcow2",if=virtio -display curses
+	qemu-system-x86_64 -monitor telnet::47023,server,nowait -machine accel=kvm,dump-guest-core=off -cpu qemu64,+avx,+avx2,+aes,+mmx,+mmxext,+popcnt,+sse,+sse2,+sse4.1,+sse4.2,+ssse3 -vga std -drive format=qcow2,file="$hdFilePath",if=virtio -device virtio-rng-pci -m size="$ram"M -smp "cores=$cores" -boot order=c -netdev user,net=192.168.47.0/24,dhcpstart=192.168.47.47,hostfwd=tcp:127.0.0.1:47022-192.168.47.47:22,id=nd1 -device virtio-net,netdev=nd1 -drive format=qcow2,file="$scratchFilePath",if=virtio -drive format=qcow2,file="$SCRIPT_DIR/admin.qcow2",if=virtio -display curses
 else
 	echo "Creating backing file of hd.qcow2..."
 	hdBackingFilePath="$scratchDir/hd_backing.qcow2"
